refactor: extract consensus spectrum generation and distance matrix computation into dedicated modules.
