+++ tests/cluster/test_consensus.py
@@ -4,7 +4,7 @@ import numpy as np
 import numba as nb
 import pytest
 
-from falcon.cluster import cluster, similarity
+from falcon.cluster import cluster, similarity, consensus
 
 
 # ---------------------------------------------------------------------------
@@ -41,7 +41,7 @@ class TestSpectrumBinning:
     def test_basic(self):
         """Two identical spectra should produce non-empty bins."""
         spectra = _make_spectrum_tuples(2)
-        bins_idx, bins_peaks, bins_mz = cluster._spectrum_binning(
+        bins_idx, bins_peaks, bins_mz = consensus._spectrum_binning(
             spectra, min_mz=50.0, max_mz=200.0, bin_size=1.0
         )
         assert len(bins_idx) > 0
@@ -68,7 +68,7 @@ class TestSpectrumBinning:
                     intensity=intensity,
                 )
             )
-        bins_idx, bins_peaks, bins_mz = cluster._spectrum_binning(
+        bins_idx, bins_peaks, bins_mz = consensus._spectrum_binning(
             spectra, min_mz=50.0, max_mz=200.0, bin_size=1.0
         )
         # The bin containing mz=150 should be filtered out (only 60% presence)
@@ -80,7 +80,7 @@ class TestSpectrumBinning:
     def test_empty_bins_excluded(self):
         """Bins with no peaks should not appear in the output."""
         spectra = _make_spectrum_tuples(3, mz_base=100.0, mz_step=1.0, n_peaks=2)
-        bins_idx, bins_peaks, bins_mz = cluster._spectrum_binning(
+        bins_idx, bins_peaks, bins_mz = consensus._spectrum_binning(
             spectra, min_mz=50.0, max_mz=200.0, bin_size=1.0
         )
         # All returned bins should have actual data
@@ -102,7 +102,7 @@ class TestOutlierRejection:
         m1 = np.array([100.0, 100.1, 99.9, 100.0], dtype=np.float32)
         m2 = np.array([101.0, 101.1, 100.9, 101.0], dtype=np.float32)
         bins_mz = nb.typed.List([m1, m2])
-        out_idx, out_peaks, out_mz = cluster._outlier_rejection(
+        out_idx, out_peaks, out_mz = consensus._outlier_rejection(
             bins_indices, bins_peaks, bins_mz, 1.5, 1.5
         )
         assert len(out_idx) == 2
@@ -114,7 +114,7 @@ class TestOutlierRejection:
         mzs = np.array([100.0, 100.1, 99.9, 100.0], dtype=np.float32)
         bins_peaks = nb.typed.List([peaks])
         bins_mz = nb.typed.List([mzs])
-        out_idx, out_peaks, out_mz = cluster._outlier_rejection(
+        out_idx, out_peaks, out_mz = consensus._outlier_rejection(
             bins_indices, bins_peaks, bins_mz, 1.5, 1.5
         )
         # The average should be closer to 0.5 (outlier 10.0 removed)
@@ -133,7 +133,7 @@ class TestSigmaClipping:
             [1.0, 1.1, 0.9, 1.0, 100.0], dtype=np.float32
         )
         mzs = np.array([100.0, 100.1, 99.9, 100.0, 100.0], dtype=np.float32)
-        result_i, result_m = cluster._sigma_clipping(intensities, mzs, 1.5, 1.5)
+        result_i, result_m = consensus._sigma_clipping(intensities, mzs, 1.5, 1.5)
         # The 100.0 outlier should be removed
         assert len(result_i) < len(intensities)
         assert np.max(result_i) < 2.0
@@ -142,7 +142,7 @@ class TestSigmaClipping:
         """When std=0, the loop should break without removing values."""
         intensities = np.array([0.5, 0.5, 0.5], dtype=np.float32)
         mzs = np.array([100.0, 100.0, 100.0], dtype=np.float32)
-        result_i, result_m = cluster._sigma_clipping(intensities, mzs, 1.5, 1.5)
+        result_i, result_m = consensus._sigma_clipping(intensities, mzs, 1.5, 1.5)
         assert len(result_i) == 3
 
 
@@ -156,7 +156,7 @@ class TestSigmaClip:
         values = np.array([1.0, 5.0, 10.0, 15.0, 20.0], dtype=np.float32)
         median = 10.0
         std = 5.0
-        mask = cluster._sigma_clip(values, median, std, 1.0, 1.0)
+        mask = consensus._sigma_clip(values, median, std, 1.0, 1.0)
         # bounds: [10 - 5, 10 + 5] = [5, 15]
         expected = np.array([False, True, True, True, False])
         np.testing.assert_array_equal(mask, expected)
@@ -171,7 +171,7 @@ class TestConstructAverageSpectrum:
         bins_indices = np.array([5, 10], dtype=np.int32)
         bins_peaks = np.array([0.5, 0.7], dtype=np.float32)
         bins_mz = np.array([100.0, 101.0], dtype=np.float32)
-        result = cluster._construct_average_spectrum(
+        result = consensus._construct_average_spectrum(
             bins_indices, bins_peaks, bins_mz,
             avg_precursor_mz=200.0, charge=2, avg_rt=60.0, cluster=5
         )
@@ -196,7 +196,7 @@ class TestGetClusterMedoids:
         rts = np.array([10.0, 20.0], dtype=np.float32)
         order_map = np.array([0, 1], dtype=np.int64)
         pdist = np.array([0.5], dtype=np.float32)  # 1 pair
-        result = cluster._get_cluster_medoids(
+        result = consensus._get_cluster_medoids(
             spectra, labels, rts, order_map, pdist
         )
         precursor_mzs, charges, mzs, intensities, retention_times, cluster_ids = result
@@ -211,7 +211,7 @@ class TestGetClusterMedoids:
         order_map = np.array([0, 1, 2], dtype=np.int64)
         # distances: (0,1)=0.1, (0,2)=0.5, (1,2)=0.2 => medoid is 1 (sum=0.3)
         pdist = np.array([0.1, 0.5, 0.2], dtype=np.float32)
-        result = cluster._get_cluster_medoids(
+        result = consensus._get_cluster_medoids(
             spectra, labels, rts, order_map, pdist
         )
         precursor_mzs, charges, mzs, intensities, retention_times, cluster_ids = result
@@ -231,7 +231,7 @@ class TestGetRepresentativeSpectra:
         rts = np.array([10.0, 20.0, 30.0], dtype=np.float32)
         order_map = np.array([0, 1, 2], dtype=np.int64)
         pdist = np.array([0.1, 0.5, 0.2], dtype=np.float32)
-        result = cluster._get_representative_spectra(
+        result = consensus._get_representative_spectra(
             spectra, labels, rts, order_map, "medoid", {"pdist": pdist}
         )
         assert len(result) == 1
@@ -243,6 +243,6 @@ class TestGetRepresentativeSpectra:
         rts = np.array([10.0, 20.0], dtype=np.float32)
         order_map = np.array([0, 1], dtype=np.int64)
         with pytest.raises(ValueError, match="Unknown consensus spectrum method"):
-            cluster._get_representative_spectra(
+            consensus._get_representative_spectra(
                 spectra, labels, rts, order_map, "nonexistent", {}
             )
